Refactor game logic into engine, ai, and persistence modules
