+++ src/game/index.js
@@ -1,0 +1,4 @@
+export * from './constants.js';
+export * from './engine.js';
+export * from './ai.js';
+export * from './statePersistence.js';
